+++ src/kill_eater_score_type.rs
@@ -1,6 +1,5 @@
 use crate::{
     Attributes,
-    AttributeValue,
     AttributeDef,
     TryFromAttributeValueU32,
     EffectType,
@@ -257,6 +256,7 @@ impl Attributes for KillEaterScoreType {
             effect_type: EffectType::Positive,
             hidden: true,
             stored_as_integer: false,
+            uses_float_value: true,
         },
         AttributeDef {
             defindex: 382,
@@ -267,6 +267,7 @@ impl Attributes for KillEaterScoreType {
             effect_type: EffectType::Positive,
             hidden: true,
             stored_as_integer: false,
+            uses_float_value: true,
         },
         AttributeDef {
             defindex: 384,
@@ -277,17 +278,13 @@ impl Attributes for KillEaterScoreType {
             effect_type: EffectType::Positive,
             hidden: true,
             stored_as_integer: false,
+            uses_float_value: true,
         },
     ];
     
-    /// Gets the attribute value.
-    fn attribute_value(&self) -> Option<AttributeValue> {
-        None
-    }
-    
     /// Gets the attribute float value.
-    fn attribute_float_value(&self) -> Option<f64> {
-        Some((*self as u32) as f64)
+    fn attribute_float_value(&self) -> Option<f32> {
+        Some((*self as u32) as f32)
     }
 }