+++ src/sheen.rs
@@ -1,7 +1,6 @@
 use crate::{
     Attribute,
     AttributeDef,
-    AttributeValue,
     DescriptionFormat,
     EffectType,
     ItemAttribute,
@@ -55,16 +54,12 @@ impl Attribute for Sheen {
         effect_type: EffectType::Positive,
         hidden: false,
         stored_as_integer: false,
+        uses_float_value: true,
     };
     
-    /// Gets the attribute value.
-    fn attribute_value(&self) -> Option<AttributeValue> {
-        None
-    }
-    
     /// Gets the attribute float value.
-    fn attribute_float_value(&self) -> Option<f64> {
-        Some((*self as u32) as f64)
+    fn attribute_float_value(&self) -> Option<f32> {
+        Some((*self as u32) as f32)
     }
 }
 
@@ -101,7 +96,7 @@ mod tests {
     
     #[test]
     fn accepts_attibute_argument() {
-        fn attribute_float_value<A: Attribute>(attribute: A) -> Option<f64> {
+        fn attribute_float_value<A: Attribute>(attribute: A) -> Option<f32> {
             attribute.attribute_float_value()
         }