+++ src/strange_part.rs
@@ -1,6 +1,5 @@
 use crate::{
     Attributes,
-    AttributeValue,
     AttributeDef,
     TryFromAttributeValueU32,
     KillEaterScoreType,
@@ -354,6 +353,7 @@ impl Attributes for StrangePart {
             effect_type: EffectType::Positive,
             hidden: true,
             stored_as_integer: false,
+            uses_float_value: true,
         },
         AttributeDef {
             defindex: 382,
@@ -364,6 +364,7 @@ impl Attributes for StrangePart {
             effect_type: EffectType::Positive,
             hidden: true,
             stored_as_integer: false,
+            uses_float_value: true,
         },
         AttributeDef {
             defindex: 384,
@@ -374,17 +375,13 @@ impl Attributes for StrangePart {
             effect_type: EffectType::Positive,
             hidden: true,
             stored_as_integer: false,
+            uses_float_value: true,
         },
     ];
     
-    /// Gets the attribute value.
-    fn attribute_value(&self) -> Option<AttributeValue> {
-        None
-    }
-    
     /// Gets the attribute float value.
-    fn attribute_float_value(&self) -> Option<f64> {
-        Some((*self as u32) as f64)
+    fn attribute_float_value(&self) -> Option<f32> {
+        Some((*self as u32) as f32)
     }
 }